feat(RUP):buscador, mostrar fsn cuando no coicide con el term.(RUP-182)
